# .github/workflows/swagger.yaml
openapi: 3.0.3
info:
  title: CargoHUB API
  description: API voor het beheren van klanten, orders, inventaris, locaties, zendingen, en transfers binnen CargoHUB.
  version: 1.0.0
servers:
  - url: https://localhost:5072/cargohub
    description: Productieomgeving
  - url: https://sandbox.api.cargohub.com
    description: Sandbox omgeving
paths:
...
  /clients/{client_id}/orders:
    get:
      summary: Haalt alle orders van een client op
      description: Haalt een lijst op van alle orders van een client
      operationId: readClientOrders
      parameters:
        - name: client_id
          in: path
          required: true
          description: ID van de klant om orders op te halen
          schema:
            type: integer
      responses:
        '200':
          description: Klant met orders gevonden
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '404':
          description: Geen klant met orders gevonden
...
  /locations/{location_id}:
    get:
      summary: Get a location by ID
      parameters:
        - name: location_id
          in: path
          required: true
          description: The ID of the location to retrieve.
          schema:
            type: integer
      responses:
        200:
          description: Location found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        404:
          description: Location not found
          content:
            text/plain:
              schema:
                type: string
                example: "No such location with Id: {location_id}"
    put:
      summary: Update a location by ID
      parameters:
        - name: location_id
          in: path
          required: true
          description: The ID of the location to update.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        200:
          description: Location updated successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Location updated successfully."
        400:
          description: Failed to update location
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to update location."
    delete:
      summary: Delete a location by ID
      parameters:
        - name: location_id
          in: path
          required: true
          description: The ID of the location to delete.
          schema:
            type: integer
      responses:
        200:
          description: Location deleted successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Location deleted successfully."
        400:
          description: Failed to delete location
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to delete location."
  /locations:
    get:
      summary: Get a list of all locations
      responses:
        200:
          description: A list of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
        404:
          description: No locations found
          content:
            text/plain:
              schema:
                type: string
                example: "No locations found"
    post:
      summary: Create a new location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        200:
          description: Location created successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Location created successfully."
        400:
          description: Failed to create location
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to create location."
  /orders/{order_id}:
    get:
      summary: Get an order by ID
      parameters:
        - name: order_id
          in: path
          required: true
          description: The ID of the order to retrieve.
          schema:
            type: integer
      responses:
        200:
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        404:
          description: Order not found
          content:
            text/plain:
              schema:
                type: string
                example: "No such order with Id: {order_id}"
    put:
      summary: Update an order by ID
      parameters:
        - name: order_id
          in: path
          required: true
          description: The ID of the order to update.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        200:
          description: Order updated successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Order updated successfully."
        400:
          description: Failed to update order
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to update order."
    delete:
      summary: Delete an order by ID
      parameters:
        - name: order_id
          in: path
          required: true
          description: The ID of the order to delete.
          schema:
            type: integer
      responses:
        200:
          description: Order deleted successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Order deleted successfully."
        400:
          description: Failed to delete order
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to delete order."
  /orders:
    get:
      summary: Get a list of all orders
      responses:
        200:
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        404:
          description: No orders found
          content:
            text/plain:
              schema:
                type: string
                example: "No orders found"
    post:
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        200:
          description: Order created successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Order created successfully."
        400:
          description: Failed to create order
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to create order."
  /orders/{order_id}/items:
    get:
      summary: Get all items in an order
      parameters:
        - name: order_id
          in: path
          required: true
          description: The ID of the order to retrieve items for.
          schema:
            type: integer
      responses:
        200:
          description: Items found in order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        404:
          description: No items found in the order
          content:
            text/plain:
              schema:
                type: string
                example: "No items found in order with Id: {order_id}"
    put:
      summary: Update items in an order
      parameters:
        - name: order_id
          in: path
          required: true
          description: The ID of the order to update items in.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PropertyItem'
      responses:
        200:
          description: Items in the order updated successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Items in order updated successfully."
        400:
          description: Failed to update items in order
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to update items in order."
...
  /shipments/{shipment_id}/orders:
    get:
      summary: Get orders of a shipment by ID
      parameters:
        - in: path
          name: shipment_id
          required: true
          schema:
            type: integer
          description: The ID of the shipment to retrieve orders for.
      responses:
        '200':
          description: Orders of the shipment found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '404':
          description: Orders not found or shipment not found
    put:
      summary: Update orders of a shipment by ID
      parameters:
        - in: path
          name: shipment_id
          required: true
          schema:
            type: integer
          description: The ID of the shipment to update orders for.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Shipment orders updated successfully
        '400':
          description: Failed to update shipment orders
...
  /warehouses/{warehouse_id}/locations:
    get:
      summary: Get locations in a specific warehouse by ID
      parameters:
        - in: path
          name: warehouse_id
          required: true
          schema:
            type: integer
          description: The ID of the warehouse to retrieve locations for.
      responses:
        '200':
          description: Locations in the warehouse found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
        '404':
          description: No locations found for the warehouse or warehouse not found
components:
  schemas:
...
    Item:
      type: object
      properties:
        uid:
          type: string
          description: Unieke ID van het item
        code:
          type: string
          description: Code van het item
        description:
          type: string
          description: Beschrijving van het item
        short_description:
          type: string
          description: Korte beschrijving van het item
        upc_code:
          type: string
          description: UPC code van het item
        model_number:
          type: string
          description: Modelnummer van het item
        commodity_code:
          type: string
          description: Commoditeit code van het item
        item_line:
          type: integer
          description: Itemlijn ID
        item_group:
          type: integer
          description: Itemgroep ID
        item_type:
          type: integer
          description: Itemtype ID
        unit_purchase_quantity:
          type: integer
          description: Aantal per aankoop eenheid
        unit_order_quantity:
          type: integer
          description: Aantal per bestel eenheid
        pack_order_quantity:
          type: integer
          description: Aantal per pack bestel eenheid
        supplier_id:
          type: integer
          description: Leverancier ID
        supplier_code:
          type: string
          description: Leverancier code
        supplier_part_number:
          type: string
          description: Leverancier onderdeelnummer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Location:
      type: object
      properties:
        id:
          type: integer
          description: Locatie ID
        warehouse_id:
          type: integer
          description: ID van het magazijn
        code:
          type: string
          description: Code van de locatie
        name:
          type: string
          description: Naam van de locatie
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Order:
      type: object
      properties:
        id:
          type: integer
          description: Order ID
        source_id:
          type: integer
          description: Bron ID
        order_date:
          type: string
          description: Datum van de bestelling
        request_date:
          type: string
          description: Verzoekdatum
        reference:
          type: string
          description: Referentie van de order
        reference_extra:
          type: string
          description: Extra referentie
        order_status:
          type: string
          description: Status van de bestelling
        notes:
          type: string
          description: Notities voor de bestelling
        shipping_notes:
          type: string
          description: Verzendnotities
        picking_notes:
          type: string
          description: Picking notities
        warehouse_id:
          type: integer
          description: Magazijn ID
        ship_to:
          type: integer
          description: Verzendadres ID
        bill_to:
          type: integer
          description: Factuuradres ID
        shipment_id:
          type: integer
          description: Zending ID
        total_amount:
          type: number
          format: float
          description: Totaalbedrag van de bestelling
        total_discount:
          type: number
          format: float
          description: Totale korting van de bestelling
        total_tax:
          type: number
          format: float
          description: Totale belasting van de bestelling
        total_surcharge:
          type: number
          format: float
          description: Totale toeslag van de bestelling
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            type: object
            properties:
              item_id:
                type: string
              quantity:
                type: integer
...
    PropertyItem:
      type: object
      properties:
        id:
          type: string
          description: Item ID
        amount:
          type: integer
          description: Amount of items